@@ -1,32 +1,4 @@
 import Router from 'next/router'
-import cookie from 'cookie'
-
-const counter = new Map()
-
-export function noop() {}
-
-export function resetCounter() {
-  counter.clear()
-}
-
-export function generateId(key) {
-  let count = 0
-
-  if (counter.has(key)) {
-    count = counter.get(key) + 1
-  }
-
-  counter.set(key, count)
-
-  return `${key}-${count}`
-}
-
-export function parseCookies(context, options = {}) {
-  return cookie.parse(
-    context.req ? context.req.headers.cookie || '' : document.cookie,
-    options,
-  )
-}
 
 export function redirect(context, target) {
   if (context.res) {
@@ -34,5 +6,37 @@ export function redirect(context, target) {
     context.res.end()
   } else {
     Router.replace(target)
+  }
+}
+
+export function wrapEvents(...handlers) {
+  return event => {
+    handlers.some(fn => {
+      if (fn) {
+        fn(event)
+      }
+      return event.defaultPrevented
+    })
+
+    return undefined
+  }
+}
+
+function assignRef(ref, value) {
+  if (ref == null) return
+  if (typeof ref === 'function') {
+    ref(value)
+  } else {
+    try {
+      ref.current = value // eslint-disable-line no-param-reassign
+    } catch (error) {
+      throw new Error(`Cannot assign value "${value}" to ref "${ref}"`)
+    }
+  }
+}
+
+export function assignRefs(...refs) {
+  return value => {
+    refs.forEach(ref => assignRef(ref, value))
   }
 }
