@@ -1,30 +1,34 @@
 import React from 'react'
-import Document, { Head, Main, NextScript } from 'next/document'
+import Document from 'next/document'
 import { ServerStyleSheet } from 'styled-components'
-import { resetCounter } from '../lib/utils'
+import { clearIds } from '../lib/useId'
 
 export default class MyDocument extends Document {
-  static getInitialProps({ renderPage }) {
+  static async getInitialProps(ctx) {
+    // Reset id counter
+    clearIds()
+
     const sheet = new ServerStyleSheet()
-    const page = renderPage(App => props =>
-      sheet.collectStyles(<App {...props} />),
-    )
-    const styleTags = sheet.getStyleElement()
+    const originalRenderPage = ctx.renderPage
 
-    resetCounter()
+    try {
+      ctx.renderPage = () =>
+        originalRenderPage({
+          enhanceApp: App => props => sheet.collectStyles(<App {...props} />),
+        })
 
-    return { ...page, styleTags }
-  }
-
-  render() {
-    return (
-      <html lang="en">
-        <Head>{this.props.styleTags}</Head>
-        <body>
-          <Main />
-          <NextScript />
-        </body>
-      </html>
-    )
+      const initialProps = await Document.getInitialProps(ctx)
+      return {
+        ...initialProps,
+        styles: (
+          <>
+            {initialProps.styles}
+            {sheet.getStyleElement()}
+          </>
+        ),
+      }
+    } finally {
+      sheet.seal()
+    }
   }
 }
